feat: Add landing page and move dashboard
