+++ middleware.js
@@ -1,0 +1,2 @@
+export { default } from "next-auth/middleware";
+export const config = { matcher: ["/dashboard"] };
